Create a new dotnet project with dotnet run which allowed me to run the program in VSCode terminal. Move code.cs into Program.cs
